Upgrade test to NET Standard style
