+++ tests/JPC.BindablePicker.Tests/BindingContext.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.ObjectModel;
+
+namespace JPC.BindablePicker.Tests
+{
+    internal class BindingContext
+    {
+        public ObservableCollection<object> Items { get; set; }
+
+        public object SelectedItem { get; set; }
+    }
+}
